Update token_splitter.py
I think this code is built around the tokeniser from tiktoken. Indeed when we call self.tokenizer(text) with a tokenizer from tiktoken we get as a result a list of tokens. However, when we use a tokenizer from the transformers library (in my case `mistralai/Mistral-7B-Instruct-v0.1`) the output of the self.tokenizer(text) is not a the list of tokens but a dictionary with two keys: input_ids and attention_mask. The input_ids contain the list of tokens as if it was generated by self.tokenizer(text) with a tiktoken tokenizer